Aggregate indexer functions into a single addIndexers function
