+++ internal/k8s/ako_init.go
@@ -32,6 +32,7 @@ import (
 	"github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/internal/rest"
 	"github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/internal/retry"
 	"github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/internal/status"
+	akov1alpha1 "github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/pkg/apis/ako/v1alpha1"
 	"github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/pkg/utils"
 	"github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/third_party/github.com/vmware/alb-sdk/go/clients"
 	"github.com/vmware/load-balancer-and-ingress-services-for-kubernetes/third_party/github.com/vmware/alb-sdk/go/session"
@@ -46,6 +47,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/cache"
+	servicesapi "sigs.k8s.io/service-apis/apis/v1alpha1"
 )
 
 func PopulateCache() error {
@@ -522,10 +524,6 @@ func (c *AviController) InitController(informers K8sinformers, registeredInforme
 
 	// Setup and start event handlers for objects.
 	c.addIndexers()
-	c.AddCrdIndexer()
-	if lib.UseServicesAPI() {
-		c.AddSvcApiIndexers()
-	}
 	c.Start(stopCh)
 
 	// once the l3 cache is populated, we can call the updatestatus functions from here
@@ -638,6 +636,7 @@ func (c *AviController) addIndexers() {
 			},
 		)
 	}
+
 	c.informers.ServiceInformer.Informer().AddIndexers(
 		cache.Indexers{
 			lib.AviSettingServicesIndex: func(obj interface{}) ([]string, error) {
@@ -654,6 +653,7 @@ func (c *AviController) addIndexers() {
 			},
 		},
 	)
+
 	if c.informers.RouteInformer != nil {
 		c.informers.RouteInformer.Informer().AddIndexers(
 			cache.Indexers{
@@ -671,6 +671,52 @@ func (c *AviController) addIndexers() {
 		)
 	}
 
+	informer := lib.AKOControlConfig().CRDInformers()
+	if lib.AKOControlConfig().AviInfraSettingEnabled() {
+		informer.AviInfraSettingInformer.Informer().AddIndexers(
+			cache.Indexers{
+				lib.SeGroupAviSettingIndex: func(obj interface{}) ([]string, error) {
+					infraSetting, ok := obj.(*akov1alpha1.AviInfraSetting)
+					if !ok {
+						return []string{}, nil
+					}
+					return []string{infraSetting.Spec.SeGroup.Name}, nil
+				},
+			},
+		)
+	}
+
+	if lib.UseServicesAPI() {
+		informer := lib.AKOControlConfig().SvcAPIInformers()
+		informer.GatewayInformer.Informer().AddIndexers(
+			cache.Indexers{
+				lib.GatewayClassGatewayIndex: func(obj interface{}) ([]string, error) {
+					gw, ok := obj.(*servicesapi.Gateway)
+					if !ok {
+						return []string{}, nil
+					}
+					return []string{gw.Spec.GatewayClassName}, nil
+				},
+			},
+		)
+
+		informer.GatewayClassInformer.Informer().AddIndexers(
+			cache.Indexers{
+				lib.AviSettingGWClassIndex: func(obj interface{}) ([]string, error) {
+					gwclass, ok := obj.(*servicesapi.GatewayClass)
+					if !ok {
+						return []string{}, nil
+					}
+					if gwclass.Spec.ParametersRef != nil {
+						// sample settingKey: ako.vmware.com/AviInfraSetting/avi-1
+						settingKey := gwclass.Spec.ParametersRef.Group + "/" + gwclass.Spec.ParametersRef.Kind + "/" + gwclass.Spec.ParametersRef.Name
+						return []string{settingKey}, nil
+					}
+					return []string{}, nil
+				},
+			},
+		)
+	}
 }
 
 func (c *AviController) FullSync() {
